22_Natural Language Processing/Lab/NER/Bi-LSTM+CRF/requirements.txt
pytorch-crf
sklearn